requirements/base.txt
django==2.0rc1
django-configurations==2.0
django-extensions==1.9.7
django-templated-mail==1.0.0
django-filter==1.1.0

graphene==2.0.1
graphene-django>=2.0.0

nameparser==0.5.3
